Module-based input notification
Module-based way to provide input feedback for user. Currently there are
only frame and none (dummy one) modules for input. Frame input module is
simple a port of the previous input indicator with few modifications and
improvements. First of all, frame can be configured (width and colors) via
the application arguments. Secondly it is drawn as a single window, which
removes unnecessary window shadows (provided by the composition manager).
However, when used with the none background module, the effect is not as
intended and this issue should be fixed.

Also main event loop has been refactored. The non-blocking version of window
event check is used only when alock is in the "typing" state, otherwise
loop is blocked until key event occurs.