Structures rewrite and initialization refactoring
Alock lacks one important feature - reinitialization upon the resolution
change of used display. Unfortunately with current application design it
was almost impossible to incorporate required changes - everything would
become one big entanglement mess.

The main purpose of this commit (and few commits before it) is to refactor
internal structures to allow "on the fly" modifications (reinitialization).
This commit also makes some simplifications to the initialization mechanism -
mainly for the maintenance and readability improvement.