Try to handle setuid errors
According to the documentation, not checking the setuid return value is a
grave security error. It is hard not to agree with this, however in our
application we have to choose between the security issue and the security
issue.

The main purpose of alock is to block unauthorized access to the X server,
so we have to omit some sub-critical errors - in this case setuid error.