Remove code supporting XFree86Misc
xf86misc has been deprecated since at least 2009. Most utilities that
depended on it (such as xset) no longer need this extension to work,
indicating that Xorg supports these things out of the box now.

I could not find a way to get xf86misc to work on a modern Xorg version,
and alock worked perfectly fine without a dependency on it, so code
supporting it was removed.

The result of this commit makes the alock offered here match what's
offered by Gentoo as of 2016-01-10.