Fix for grayscale conversion
The grayscale conversion depends on the depth of the image and the image
depth - in our case - depends on the depth of the screen. In most cases,
screen depth is set to 24 bits, however this assumption leads to weird
glitches...

The new code has explicit branches for different types of depths. Also the
conversion is done according to the colorimetric strategy - the grayscale
image is more GIMP-like.