Customize build-in default PAM service
It seems, that there is no sane default for build-in PAM service. In
order to simplify package maintenance for different Linux distributions,
this commit allows specifying default PAM service by the package
maintainer.

Note, that this commit makes non-backward compatible change. The PAM
"module" is renamed to PAM "service", which is consistent with the PAM
documentation.

Fixes #5